# services/goldfin-admin-server/src/gen/java/io/goldfin/admin/service/api/model/DocumentParameters.java
/*
 * Goldfin Service API
 * REST API for Goldfin Intelligent Invoice Processing
 *
 * OpenAPI spec version: 1.0.0
 * Contact: info@goldfin.io
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


package io.goldfin.admin.service.api.model;

import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import io.goldfin.admin.service.api.model.TagSet;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import javax.validation.constraints.*;

/**
 * User settable document parameters
 */
@ApiModel(description = "User settable document parameters")

public class DocumentParameters   {
  @JsonProperty("description")
  private String description = null;

  @JsonProperty("tags")
  private TagSet tags = null;

  public DocumentParameters description(String description) {
    this.description = description;
    return this;
  }

  /**
   * Optional description of the document
   * @return description
   **/
  @JsonProperty("description")
  @ApiModelProperty(value = "Optional description of the document")
  public String getDescription() {
    return description;
  }

  public void setDescription(String description) {
    this.description = description;
  }

  public DocumentParameters tags(TagSet tags) {
    this.tags = tags;
    return this;
  }

  /**
   * Get tags
   * @return tags
   **/
  @JsonProperty("tags")
  @ApiModelProperty(value = "")
  public TagSet getTags() {
    return tags;
  }

  public void setTags(TagSet tags) {
    this.tags = tags;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    DocumentParameters documentParameters = (DocumentParameters) o;
    return Objects.equals(this.description, documentParameters.description) &&
        Objects.equals(this.tags, documentParameters.tags);
  }

  @Override
  public int hashCode() {
    return Objects.hash(description, tags);
  }


  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class DocumentParameters {\n");
    
    sb.append("    description: ").append(toIndentedString(description)).append("\n");
    sb.append("    tags: ").append(toIndentedString(tags)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}