# swagger/goldfin-service-admin-api.yaml
swagger: '2.0'
info:
  description: REST API for Goldfin Intelligent Invoice Processing
  version: 1.0.0
  title: Goldfin Service API
  termsOfService: 'http://goldfin.io/terms/'
  contact:
    email: info@goldfin.io
host: api.goldfin.io
basePath: /api/v1
...
schemes:
  - https
securityDefinitions:
  SessionKey:
    type: apiKey
    in: header
    name: vnd.io.goldfin.session
  ApiKey:
    type: apiKey
    in: header
    name: vnd.io.goldfin.apikey
security:
  - SessionKey: []
  - ApiKey: []
paths:
...
  /document:
    post:
      tags:
        - document
      summary: Upload document
      description: Upload a new document for scanning
      operationId: documentCreate
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - name: description
          in: formData
          description: An optional description of the document
          type: string
        - name: tags
          in: formData
          description: Optional tags that apply to this document passed as a JSON string containing name-value pairs.
          type: string
        - name: process
          in: formData
          description: Optional flag to kick off scanning automatically if true
          type: boolean
          default: true
        - name: file
          in: formData
          description: Document file
          required: true
          type: file
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/Document'
        '400':
          description: Invoice creation failed
          schema:
            $ref: '#/definitions/ApiResponse'
    get:
      tags:
        - document
      summary: List documents
      description: Return a list of all documents
      operationId: documentShowAll
      produces:
        - application/json
      responses:
        '200':
          description: Successful query
          schema:
            type: array
            items:
              $ref: '#/definitions/Document'
  '/document/{id}':
    get:
      tags:
        - document
      summary: Return document metadata
      description: Download document metadata without content
      operationId: documentShow
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: Document ID
          required: true
          type: string
      responses:
        '200':
          description: Successful
          schema:
            $ref: '#/definitions/Document'
        '404':
          description: Not Found
    put:
      tags:
        - document
      summary: Update a document
      description: Update document description and/or tags. Tags do not propagate to already scanned invoices but will apply to new ones. 
      operationId: documentUpdate
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          description: Document ID
          required: true
          type: string
        - name: body
          in: body
          description: Document parameters
          schema:
            $ref: '#/definitions/DocumentParameters'
      responses:
        '200':
          description: Successful
        '404':
          description: Not Found
    delete:
      tags:
        - document
      summary: Delete a invoice
      description: Delete a single document and associated semantic content
      operationId: documentDelete
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: Invoice ID
          required: true
          type: string
      responses:
        '200':
          description: Successful
        '404':
          description: Not Found
  '/document/{id}/download':
    get:
      tags:
        - document
      summary: Download content
      description: Download document content
      operationId: documentDownload
      produces:
        - application/pdf
        - application/octet-stream
      parameters:
        - name: id
          in: path
          description: Document ID
          required: true
          type: string
      responses:
        '200':
          description: Successful query
          schema:
            type: file
        '404':
          description: Not Found
  '/document/{id}/process':
    post:
      tags:
        - document
      summary: Kick off document analysis
      description: >-
        Run background scanning on document.  The document state and semantic
        information will be updated when finished.
      operationId: documentProcess
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: Document ID
          required: true
          type: string
      responses:
        '202':
          description: Accepted
        '404':
          description: Not Found
...
definitions:
...
  Document:
    description: 'A document, for example a PDF file containing an invoice'
    type: object
    properties:
      id:
        type: string
        description: Document ID
        format: uuid
      name:
        type: string
        description: Name of the document
      description:
        type: string
        description: Optional description of the document
      contentType:
        type: string
        description: 'Internet media type (e.g., application/octet-stream)'
      contentLength:
        type: number
        format: int64
        description: Document length in bytes
      thumbprint:
        type: string
        description: SHA-256 thumbprint of document content
      locator:
        type: string
        format: url
        description: Storage locator of the document
      state:
        description: >-
          The current processing state of the document.  The document type and
          content ID are available after scanning
        enum:
          - CREATED
          - SCAN_REQUESTED
          - SCANNED
          - ERROR
      semanticType:
        description: 'Kind of document, e.g., an invoice.'
        enum:
          - INVOICE
          - UNKNOWN
      semanticId:
        type: string
        description: 'ID of the document''s scanned content, e.g. an invoice ID'
        format: uuid
      tags:
        $ref: '#/definitions/TagSet'
      creationDate:
        type: string
        description: Date document was uploaded
  DocumentParameters:
    description: User settable document parameters
    type: object
    properties:
      description:
        type: string
        description: Optional description of the document
      tags:
        $ref: '#/definitions/TagSet'
...
  TagSet:
    type: array
    description: Collection of tags.  If keys are repeated in the array the last value in order takes precedences
    items:
      $ref: '#/definitions/Tag'
  Tag:
    type: object
    description: A key-value pair used to add relevant descriptive information. Tags are not interpreted and may contain data like business unit IDs, contract names, or other information users find relevant to classify or associate the data to which they apply. 
    properties:
      name:
        type: string
        description: A user-defined name for the tag
      value:
        type: string
        description: A user-defined value for the tag
  ApiResponse:
    type: object
    properties:
      code:
        type: integer
        format: int32
      type:
        type: string
      message:
        type: string
...